requirements.txt
streamlit==1.31.0
pandas==2.0.3
matplotlib==3.7.2
numpy==1.24.3
PyPDF2==3.0.1
python-docx==0.8.11
python-dateutil==2.8.2
google-generativeai==0.3.2
protobuf==4.25.1
seaborn==0.12.2
pillow==10.0.0
sentence_transformers==2.2.2  #  Crucial: Add this, and a specific version
huggingface_hub==0.16.4       #   And this with a specific, compatible version